Refactors visitor and qr express routes #127
