+++ react-backend/routes/qr_generator.js
@@ -1,43 +1,22 @@
-const express = require('express');
-const putUserData = require('../database/queries/putFormData');
+const router = require('express').Router();
+const userRegister = require('../database/queries/user_register');
 const qrcodemaker = require('../functions/qrcodemaker');
 const hash = require('../functions/hash');
-const sendQrCode = require('./qr_send');
-
-const router = express.Router();
+const sendQrCode = require('../functions/qr_send');
 
 router.post('/', (req, res, next) => {
-  let body = '';
-  req.on('data', chunk => {
-    body += chunk;
-  });
+  const { formSender, formSex, formYear, formEmail } = req.body;
+  const hashString = hash(req.body);
+  const name = formSender.toLowerCase();
 
-  req.on('end', () => {
-    const details = JSON.parse(body);
-    const hashString = hash(details);
-    const name = details.formSender.toLowerCase();
-
-    putUserData(
-      req.auth.cb_id,
-      name,
-      details.formSex,
-      details.formYear,
-      details.formEmail,
-      hashString,
-    )
-      .then(() => {
-        sendQrCode(
-          details.formEmail,
-          details.formSender,
-          hashString,
-          req.auth.cb_logo,
-        );
-        return hashString;
-      })
-      .then(qrcodemaker)
-      .then(qr => res.send({ qr, cb_logo: req.auth.cb_logo }))
-      .catch(next);
-  });
+  userRegister(req.auth.cb_id, name, formSex, formYear, formEmail, hashString)
+    .then(() => {
+      sendQrCode(formEmail, formSender, hashString, req.auth.cb_logo);
+      return hashString;
+    })
+    .then(qrcodemaker)
+    .then(qr => res.send({ qr, cb_logo: req.auth.cb_logo }))
+    .catch(next);
 });
 
 module.exports = router;
