+++ react-backend/routes/qr_user_gen.js
@@ -3,9 +3,7 @@ const qrCodeMaker = require('../functions/qrcodemaker');
 
 router.post('/', (req, res, next) => {
   qrCodeMaker(req.body.hash)
-    .then(qr =>
-      res.send({ token: req.auth.adminToken, qr, cb_logo: req.auth.cb_logo })
-    )
+    .then(qr => res.send({ qr, cb_logo: req.auth.cb_logo }))
     .catch(next);
 });
