--- react-backend/routes/visitor_activity.js
@@ -1,20 +1,0 @@
-const express = require('express');
-const putVisitsData = require('../database/queries/putVisitsData');
-
-const router = express.Router();
-
-router.post('/', (req, res, next) => {
-  let body = '';
-  req.on('data', chunk => {
-    body += chunk;
-  });
-
-  req.on('end', () => {
-    const visitToAdd = JSON.parse(body);
-    putVisitsData(visitToAdd.hash, visitToAdd.activity, req.auth.cb_id)
-      .then(() => res.send('success'))
-      .catch(next);
-  });
-});
-
-module.exports = router;
